Refine LGT Java interface inheritance
